Fix issues and performance for data transmission
Combine eeg and acc in a single transaction for lower load and fix a race issue
Upload by chunk even if finished until all data is uploaded
Make sure hte service is restarted correctly

+++ core/src/main/java/io/nextsense/android/base/db/objectbox/ObjectBoxDatabase.java
@@ -152,7 +152,7 @@ public class ObjectBoxDatabase implements Database {
       accelerationQuery.setParameter(Acceleration_.localSessionId, localSessionId).find());
   }
 
-  public List<EegSample> getLastEegSamples(int localSessionId, long count) {
+  public List<EegSample> getLastEegSamples(long localSessionId, long count) {
     return runWithExceptionLog(() -> {
       long sessionEegSamplesCount = getEegSamplesCount(localSessionId);
       long offset = count <= sessionEegSamplesCount ? sessionEegSamplesCount - count : 0;
